adds Github Oath
